Move mock data into __mocks__
