+++ src/test/__mocks__/applicationTypesData.js
