+++ src/test/components/SourceRemoveModal.spec.js
@@ -10,8 +10,8 @@ import { MemoryRouter } from 'react-router-dom';
 import * as actions from '../../redux/sources/actions';
 import SourceRemoveModal from '../../components/SourceRemoveModal';
 import { componentWrapperIntl } from '../../utilities/testsHelpers';
-import { sourcesDataGraphQl } from '../sourcesData';
-import { applicationTypesData } from '../applicationTypesData';
+import { sourcesDataGraphQl } from '../__mocks__/sourcesData';
+import { applicationTypesData } from '../__mocks__/applicationTypesData';
 import RemoveAppModal from '../../components/AddApplication/RemoveAppModal';
 import ApplicationList from '../../components/ApplicationsList/ApplicationList';
 import { routes, replaceRouteId } from '../../Routes';
