+++ src/test/components/SourcesSimpleView/formatters.spec.js
@@ -24,8 +24,8 @@ import {
     UNAVAILABLE,
     UnknownError
 } from '../../../components/SourcesSimpleView/formatters';
-import { sourceTypesData, OPENSHIFT_ID, AMAZON_ID, OPENSHIFT_INDEX } from '../../sourceTypesData';
-import { sourcesDataGraphQl, SOURCE_CATALOGAPP_INDEX, SOURCE_ALL_APS_INDEX, SOURCE_NO_APS_INDEX, SOURCE_ENDPOINT_URL_INDEX } from '../../sourcesData';
+import { sourceTypesData, OPENSHIFT_ID, AMAZON_ID, OPENSHIFT_INDEX } from '../../__mocks__/sourceTypesData';
+import { sourcesDataGraphQl, SOURCE_CATALOGAPP_INDEX, SOURCE_ALL_APS_INDEX, SOURCE_NO_APS_INDEX, SOURCE_ENDPOINT_URL_INDEX } from '../../__mocks__/sourcesData';
 import {
     applicationTypesData,
     CATALOG_INDEX,
@@ -33,7 +33,7 @@ import {
     COSTMANAGEMENET_INDEX,
     COSTMANAGEMENT_APP,
     CATALOG_APP
-} from '../../applicationTypesData';
+} from '../../__mocks__/applicationTypesData';
 import { Badge, Tooltip, Bullseye } from '@patternfly/react-core';
 import { DateFormat } from '@redhat-cloud-services/frontend-components/components/DateFormat';
 import { IntlProvider } from 'react-intl';
