+++ src/test/components/sourceEditForm/SourceEditModal.spec.js
@@ -10,9 +10,9 @@ import { Spinner } from '@patternfly/react-core/dist/js/components/Spinner';
 import { componentWrapperIntl } from '../../../utilities/testsHelpers';
 import SourceEditModal from '../../../components/SourceEditForm/SourceEditModal';
 import { routes, replaceRouteId } from '../../../Routes';
-import { applicationTypesData } from '../../applicationTypesData';
-import { sourceTypesData, OPENSHIFT_ID } from '../../sourceTypesData';
-import { sourcesDataGraphQl } from '../../sourcesData';
+import { applicationTypesData } from '../../__mocks__/applicationTypesData';
+import { sourceTypesData, OPENSHIFT_ID } from '../../__mocks__/sourceTypesData';
+import { sourcesDataGraphQl } from '../../__mocks__/sourcesData';
 import { Modal, Button, FormGroup, TextInput, Form } from '@patternfly/react-core';
 import * as editApi from '../../../api/doLoadSourceForEdit';
 import EditFieldProvider from '../../../components/editField/EditField';
